Reworked agent list
The AGENTS array is now empty by default, and is filled in with the readAgentsFile function to add agents to the program. This allows the end user to now manually choose which agents are/aren't allowed to be picked, as well as make it a bit more streamlined to add new agents to the program by the end user.